+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 * Tools: added a new "flythrough" camera mode for navigating viewports.
 * Tools: added textual tooltips to main object types and prominent interface elements.
 * Tools: crown-launcher can now be executed from any directory.
+* Data Compile: resource packages are now LZ4-compressed.
 
 0.60.0 --- 30 Dec 2025
 ----------------------